Fix ehphp interface

+++ ehi/make_ehphp.sh
@@ -5,8 +5,10 @@ OS=`uname`
 case $OS in
 	'Darwin')
 		$1 -dynamiclib $2 -o ehphp.dylib $3
-		sudo cp ehphp.dylib /opt/local/lib/php54/extensions/no-debug-non-zts-20100525/ehphp.dylib
-		sudo cp ehphp.dylib /opt/local/lib/php54/extensions/debug-non-zts-20100525/ehphp.dylib
+		if [ -e /opt/local/lib/php54/extensions/no-debug-non-zts-20100525 ]; then
+			sudo cp ehphp.dylib /opt/local/lib/php54/extensions/no-debug-non-zts-20100525/ehphp.dylib
+			sudo cp ehphp.dylib /opt/local/lib/php54/extensions/debug-non-zts-20100525/ehphp.dylib
+		fi
 		;;
 	'Linux')
 		$1 -shared -o ehphp.so $3
